Move assertEqual/assertNotEqual to internal/testutils
