+++ otel/helpers_test.go
@@ -7,64 +7,21 @@ package sentryotel
 
 import (
 	"encoding/hex"
-	"fmt"
-	"reflect"
 	"sort"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/getsentry/sentry-go"
+	"github.com/getsentry/sentry-go/internal/testutils"
 	"github.com/google/go-cmp/cmp"
 	"go.opentelemetry.io/otel/baggage"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/trace"
 )
 
-func assertEqual(t *testing.T, got, want interface{}, userMessage ...interface{}) {
-	t.Helper()
-
-	if !reflect.DeepEqual(got, want) {
-		logFailedAssertion(t, formatUnequalValues(got, want), userMessage...)
-	}
-}
-
-func assertNotEqual(t *testing.T, got, want interface{}, userMessage ...interface{}) {
-	t.Helper()
-
-	if reflect.DeepEqual(got, want) {
-		logFailedAssertion(t, formatUnequalValues(got, want), userMessage...)
-	}
-}
-
-func logFailedAssertion(t *testing.T, summary string, userMessage ...interface{}) {
-	t.Helper()
-	text := summary
-
-	if len(userMessage) > 0 {
-		if message, ok := userMessage[0].(string); ok {
-			if message != "" && len(userMessage) > 1 {
-				text = fmt.Sprintf(message, userMessage[1:]...) + text
-			} else if message != "" {
-				text = fmt.Sprint(message) + text
-			}
-		}
-	}
-
-	t.Error(text)
-}
-
-func formatUnequalValues(got, want interface{}) string {
-	var a, b string
-
-	if reflect.TypeOf(got) != reflect.TypeOf(want) {
-		a, b = fmt.Sprintf("%T(%#v)", got, got), fmt.Sprintf("%T(%#v)", want, want)
-	} else {
-		a, b = fmt.Sprintf("%#v", got), fmt.Sprintf("%#v", want)
-	}
-
-	return fmt.Sprintf("\ngot: %s\nwant: %s", a, b)
-}
+var assertEqual = testutils.AssertEqual
+var assertNotEqual = testutils.AssertNotEqual
 
 // assertMapCarrierEqual compares two values of type propagation.MapCarrier and raises an
 // assertion error if the values differ.
@@ -159,7 +116,7 @@ func otelSpanIDFromHex(s string) trace.SpanID {
 	return spanID
 }
 
-// FIXME(anton): copie from mocks_test.go
+// FIXME(anton): copied from mocks_test.go
 
 type TransportMock struct {
 	mu        sync.Mutex
